+++ dist-esnext/BaseRepository.js
@@ -1,10 +1,12 @@
+import { MongoClient } from 'mongodb';
 import R from 'ramda';
-const getCollection = R.curry((db, collectionName) => db.collection(collectionName));
+const getDb = async (url) => await MongoClient.connect(url);
+const getDbCollection = R.curry((db, collectionName) => db.collection(collectionName));
 const save = R.curry(async (collection, entity) => {
     const result = await collection
         .replaceOne({ _id: entity.id }, entity, { upsert: true });
     entity = result.ops[0];
-    return Promise.resolve(entity);
+    return entity;
 });
 const getById = R.curry((collection, id) => {
     const query = {
@@ -24,5 +26,5 @@ const getByIds = R.curry((collection, ids) => {
         .toArray();
 });
 const find = R.curry((collection, query, options) => collection.find(query, {}, options).toArray());
-export { save, find, getCollection, getById, getByIds };
+export { save, find, getDb, getDbCollection, getById, getByIds };
 //# sourceMappingURL=BaseRepository.js.map