+++ src/BaseRepository.ts
@@ -1,20 +1,24 @@
-import { Collection, Db, WriteOpResult } from 'mongodb';
+import { Collection, Db, MongoClient, WriteOpResult } from 'mongodb';
 import { IEntityMinBase } from 'ptz-core-domain';
 import R from 'ramda';
 
-const getCollection = R.curry((db: Db, collectionName: string) =>
-    db.collection<IEntityMinBase>(collectionName)
+const getDb = async (url: string) => await MongoClient.connect(url);
+
+const getDbCollection = R.curry((db: Db, collectionName: string) =>
+    db.collection(collectionName)
 );
 
 const save = R.curry(async (collection: Collection | any, entity: IEntityMinBase) => {
     const result = await collection
-        .replaceOne({ _id: entity.id }, entity, { upsert: true }) as WriteOpResult;
+        .replaceOne(
+        { _id: entity.id },
+        entity, { upsert: true }) as WriteOpResult;
 
     entity = result.ops[0];
-    return Promise.resolve(entity);
+    return entity;
 });
 
-const getById = R.curry((collection: Collection<IEntityMinBase>, id: string) => {
+const getById = R.curry((collection: Collection, id: string) => {
     const query = {
         _id: id
     };
@@ -23,7 +27,7 @@ const getById = R.curry((collection: Collection<IEntityMinBase>, id: string) => 
         .findOne(query);
 });
 
-const getByIds = R.curry((collection: Collection<IEntityMinBase>, ids: string[]) => {
+const getByIds = R.curry((collection: Collection, ids: string[]) => {
     const query = {
         _id: {
             $in: ids
@@ -35,8 +39,8 @@ const getByIds = R.curry((collection: Collection<IEntityMinBase>, ids: string[])
         .toArray();
 });
 
-const find = R.curry((collection: Collection<IEntityMinBase> | any, query: any, options: { limit: number }) =>
+const find = R.curry((collection: Collection | any, query: any, options: { limit: number }) =>
     collection.find(query, {}, options).toArray()
 );
 
-export { save, find, getCollection, getById, getByIds };
+export { save, find, getDb, getDbCollection, getById, getByIds };